requirements.txt
streamlit
pandas
reportlab
arabic-reshaper
python-bidi
pydrive2


